MC Web: move sandbox fleet module out of github/ into src/web/sandbox/
Sandbox provisioning (derive, budget, reattach, workdir layout) is not
GitHub-specific, so it now lives in its own module. GitHub-specific
materialization/worktree/PR helpers stay in github/sandbox.ts.
