+++ mastracode/web/src/web/github/config.ts
@@ -11,7 +11,7 @@ import { createHmac, randomBytes, timingSafeEqual } from 'node:crypto';
 import { isWebAuthEnabled } from '../auth';
 import { getMissingGithubAppEnvVars, isGithubAppConfigured } from './client';
 import { isAppDbConfigured } from './db';
-import { getSandboxProvider, isSandboxEnabled } from './sandbox';
+import { getSandboxProvider, isSandboxEnabled } from '../sandbox/fleet';
 
 /**
  * True when the GitHub App project feature should be active.
