+++ mastracode/web/src/web/sandbox-reattach-registration.ts
@@ -6,8 +6,8 @@
  * is registered here at web-surface load time.
  */
 import { registerSandboxReattach as registerOnCore } from '@mastra/code-sdk/agents/sandbox-reattach';
-import { reattachProjectSandbox } from './github/sandbox.js';
+import { reattachSandbox } from './sandbox/fleet.js';
 
 export function registerSandboxReattach(): void {
-  registerOnCore(reattachProjectSandbox);
+  registerOnCore(reattachSandbox);
 }
